aria/requirements.txt
# ARIA — Python dependencies
# Install with:  pip install -r requirements.txt

# LLM / agent framework
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.0

# Web search (no API key required)
duckduckgo-search>=6.1.0

# HTTP client (used by SerpAPI fallback)
requests>=2.32.0
